+++ package.json
@@ -4,8 +4,8 @@
   "description": "Blog style web page for the team Almighty Lobsters",
   "main": "server.js",
   "scripts": {
-    "start": "node index.js",
-    "start:dev": "nodemon index.js"
+    "start": "node server.js",
+    "start:dev": "nodemon server.js"
   },
   "repository": {
     "type": "git",
