+++ web.config
@@ -7,16 +7,5 @@
 			<mimeMap fileExtension=".woff" mimeType="application/font-woff" />
 			<mimeMap fileExtension=".woff2" mimeType="application/font-woff2" />
 		</staticContent>
-		<hanlers>
-			<add name="iisnode" path="index.js" verb="*" modules="iisnode" />
-		</hanlers>
-		<rewrite>
-			<rules>
-				<rule name="DynamicContent">
-					<match url="/*" />
-					<action type="Rewrite" url="index.js" />
-				</rule>
-			</rules>
-		</rewrite>
 	</system.webServer>
 </configuration>
